requirements.txt

Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
asgiref==3.6.0
backports.zoneinfo==0.2.1
psycopg2-binary==2.9.6
sqlparse==0.4.3



